Feature: allow specifying semver to be optional
A common flow is to use a develop branch to allow commits to stack
before merging into master; in fact we're using this in GitSavvy,
and every time we don't select a semver maintainerd complains.

For this sort of flow, I believe it should be allowed to toggle
whether selecting at least 1 semver checkbox is required or not.